services/django_backend/requirements.txt
asgiref==3.8.1
astroid==3.3.6
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dill==0.3.9
Django==5.1.4
django-filter==24.3
django-postgres-extensions==0.9.3
djangorestframework==3.15.2
Faker==33.1.0
fonttools==4.55.3
idna==3.10
isort==5.13.2
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.10.0
mccabe==0.7.0
mysqlclient==2.2.6
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.5
pycparser==2.22
PyJWT==2.10.1
pylint==3.3.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
